[refactor] Fix package misspelling: telemetery->telemetry
