+++ cmd/jaeger/config-cassandra.yaml
@@ -13,7 +13,7 @@ service:
       address: 0.0.0.0:8888
     logs:
       level: debug
-    # TODO Initialize telemetery tracer once OTEL released new feature.
+    # TODO Initialize telemetry tracer once OTEL released new feature.
     # https://github.com/open-telemetry/opentelemetry-collector/issues/10663
 
 extensions:
