+++ cmd/jaeger/internal/all-in-one.yaml
@@ -13,7 +13,7 @@ service:
       address: "${env:JAEGER_LISTEN_HOST:-localhost}:8888"
     logs:
       level: info
-    # TODO Initialize telemetery tracer once OTEL released new feature.
+    # TODO Initialize telemetry tracer once OTEL released new feature.
     # https://github.com/open-telemetry/opentelemetry-collector/issues/10663
 
 extensions:
