+++ cmd/remote-storage/app/server_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/grpctest"
 	"github.com/jaegertracing/jaeger/pkg/config/tlscfg"
 	"github.com/jaegertracing/jaeger/pkg/healthcheck"
-	"github.com/jaegertracing/jaeger/pkg/telemetery"
+	"github.com/jaegertracing/jaeger/pkg/telemetry"
 	"github.com/jaegertracing/jaeger/pkg/tenancy"
 	"github.com/jaegertracing/jaeger/ports"
 	"github.com/jaegertracing/jaeger/proto-gen/storage_v1"
@@ -45,7 +45,7 @@ func TestNewServer_CreateStorageErrors(t *testing.T) {
 	factory.On("CreateSpanWriter").Return(nil, nil)
 	factory.On("CreateDependencyReader").Return(nil, errors.New("no deps")).Once()
 	factory.On("CreateDependencyReader").Return(nil, nil)
-	telset := telemetery.Setting{
+	telset := telemetry.Setting{
 		Logger:       zap.NewNop(),
 		ReportStatus: func(*componentstatus.Event) {},
 		LeveledMeterProvider: func(_ configtelemetry.Level) metric.MeterProvider {
@@ -119,9 +119,9 @@ func TestNewServer_TLSConfigError(t *testing.T) {
 		KeyPath:      "invalid/path",
 		ClientCAPath: "invalid/path",
 	}
-	telset := telemetery.Setting{
+	telset := telemetry.Setting{
 		Logger:       zap.NewNop(),
-		ReportStatus: telemetery.HCAdapter(healthcheck.New()),
+		ReportStatus: telemetry.HCAdapter(healthcheck.New()),
 	}
 	storageMocks := newStorageMocks()
 	_, err := NewServer(
@@ -327,9 +327,9 @@ func TestServerGRPCTLS(t *testing.T) {
 			storageMocks.reader.On("GetServices", mock.AnythingOfType("*context.valueCtx")).Return(expectedServices, nil)
 
 			tm := tenancy.NewManager(&tenancy.Options{Enabled: true})
-			telset := telemetery.Setting{
+			telset := telemetry.Setting{
 				Logger:       flagsSvc.Logger,
-				ReportStatus: telemetery.HCAdapter(flagsSvc.HC()),
+				ReportStatus: telemetry.HCAdapter(flagsSvc.HC()),
 			}
 			server, err := NewServer(
 				serverOptions,
@@ -376,9 +376,9 @@ func TestServerHandlesPortZero(t *testing.T) {
 	zapCore, logs := observer.New(zap.InfoLevel)
 	flagsSvc.Logger = zap.New(zapCore)
 	storageMocks := newStorageMocks()
-	telset := telemetery.Setting{
+	telset := telemetry.Setting{
 		Logger:       flagsSvc.Logger,
-		ReportStatus: telemetery.HCAdapter(flagsSvc.HC()),
+		ReportStatus: telemetry.HCAdapter(flagsSvc.HC()),
 	}
 	server, err := NewServer(
 		&Options{GRPCHostPort: ":0"},
