+++ pkg/telemetry/settings.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package telemetery
+package telemetry
 
 import (
 	"go.opentelemetry.io/collector/component"
